+++ testing/tests/DevExpress.ui.widgets.scheduler/integration.appointmentsVertical.tests.js
@@ -1,4 +1,4 @@
-var $ = require("jquery");
+import $ from "jquery";
 
 QUnit.testStart(function() {
     $("#qunit-fixture").html(
@@ -7,36 +7,45 @@ QUnit.testStart(function() {
             </div>');
 });
 
-require("common.css!");
-require("generic_light.css!");
+import "common.css!";
+import "generic_light.css!";
 
+import "ui/scheduler/ui.scheduler";
+import "ui/switch";
 
-var translator = require("animation/translator"),
-    fx = require("animation/fx"),
-    pointerMock = require("../../helpers/pointerMock.js"),
-    Color = require("color"),
-    devices = require("core/devices"),
-    dragEvents = require("events/drag"),
-    DataSource = require("data/data_source/data_source").DataSource,
-    subscribes = require("ui/scheduler/ui.scheduler.subscribes"),
-    dataUtils = require("core/element_data");
+import { SchedulerTestWrapper } from './helpers.js';
 
-require("ui/scheduler/ui.scheduler");
-require("ui/switch");
+const createInstance = function(options) {
+    const defaultOption = {
+        maxAppointmentsPerCell: null
+    };
+    const instance = $("#scheduler").dxScheduler($.extend(defaultOption, options)).dxScheduler("instance");
+    return new SchedulerTestWrapper(instance);
+};
 
-var DATE_TABLE_CELL_CLASS = "dx-scheduler-date-table-cell",
-    APPOINTMENT_CLASS = "dx-scheduler-appointment";
+import translator from "animation/translator";
+import fx from "animation/fx";
+import pointerMock from "../../helpers/pointerMock.js";
+import Color from "color";
+import devices from "core/devices";
+import dragEvents from "events/drag";
+import { DataSource } from "data/data_source/data_source";
+import subscribes from "ui/scheduler/ui.scheduler.subscribes";
+import dataUtils from "core/element_data";
 
-var APPOINTMENT_DEFAULT_OFFSET = 25,
-    APPOINTMENT_MOBILE_OFFSET = 50;
+const DATE_TABLE_CELL_CLASS = "dx-scheduler-date-table-cell";
+const APPOINTMENT_CLASS = "dx-scheduler-appointment";
 
-function getOffset() {
+var APPOINTMENT_DEFAULT_OFFSET = 25;
+const APPOINTMENT_MOBILE_OFFSET = 50;
+
+const getOffset = () => {
     if(devices.current().deviceType !== "desktop") {
         return APPOINTMENT_MOBILE_OFFSET;
     } else {
         return APPOINTMENT_DEFAULT_OFFSET;
     }
-}
+};
 
 QUnit.module("Integration: Appointments on vertical views (day, week, workWeek)", {
     beforeEach: function() {
@@ -1505,3 +1514,28 @@ QUnit.test("Long appointments should be rendered correctly in vertical grouped w
     assert.roughEqual($appointments.eq(1).position().left, cellWidth * 2 + dateTableLeftOffset, 1.1, "correct left position of appointment part");
 });
 
+QUnit.test("Scheduler recurrent appointments render right if began before startDayHour (T735635)", function(assert) {
+    let scheduler = createInstance();
+    const appointments = [
+        {
+            text: "Website Re-Design Plan",
+            startDate: new Date(2019, 3, 22, 7, 30),
+            endDate: new Date(2019, 3, 22, 11, 30),
+            recurrenceRule: "FREQ=DAILY"
+        }
+    ];
+    const options = {
+        dataSource: appointments,
+        views: ["week", "month"],
+        currentView: "week",
+        startDayHour: 10,
+        endDayHour: 16,
+        height: 600,
+        currentDate: new Date(2019, 3, 21),
+    };
+    scheduler.instance.option(options);
+
+    const initialAppointmentHeight = scheduler.appointments.getAppointmentHeight(0);
+    const recurrentAppointmentHeight = scheduler.appointments.getAppointmentHeight(1);
+    assert.equal(initialAppointmentHeight, recurrentAppointmentHeight, "Appointment cells have equal heights (both initial and recurrent)");
+});
